remove gleitende spitzenstunde from persisting
