+++ src/main/java/de/muenchen/dave/services/processzaehldaten/ProcessZaehldatenBelastungsplanService.java
@@ -21,7 +21,7 @@ import de.muenchen.dave.repositories.elasticsearch.ZaehlstelleIndex;
 import de.muenchen.dave.repositories.relationaldb.ZeitintervallRepository;
 import de.muenchen.dave.services.ladezaehldaten.LadeZaehldatenService;
 import de.muenchen.dave.util.CalculationUtil;
-import de.muenchen.dave.util.dataimport.ZeitintervallGleitendeSpitzenstundeUtilNg;
+import de.muenchen.dave.util.dataimport.ZeitintervallGleitendeSpitzenstundeUtil;
 import de.muenchen.dave.util.dataimport.ZeitintervallSortingIndexUtil;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -729,7 +729,7 @@ public class ProcessZaehldatenBelastungsplanService {
                             spitzenstunde.getEndeUhrzeit(),
                             Set.of(options.getIntervall().getTypeZeitintervall()));
 
-            return ZeitintervallGleitendeSpitzenstundeUtilNg
+            return ZeitintervallGleitendeSpitzenstundeUtil
                     .getGleitendeSpitzenstundenByBewegungsbeziehung(
                             UUID.fromString(zaehlung.getId()),
                             options.getZeitblock(),
